% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naq.detect.R
\name{naq.detect}
\alias{naq.detect}
\title{Move raw data from an 'inbox' directory to a repository.}
\usage{
naq.detect()
}
\value{
If multiple USB volumes are detected, throws an error message.
}
\description{
Move raw data from an 'inbox' directory to a repository.
}
